requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
boto3==1.34.147
botocore==1.34.147
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.6.2
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
fastavro==1.9.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.2
idna==3.7
instructor==1.3.6
jiter==0.4.2
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.37.0
packaging==24.1
parameterized==0.9.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
